refactor(snowflake): merged snowflakeids connector with existing snowflake connector
